+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.0.3] - 2019-03-21
 - Fix for Fotorama gallery
 - Check via timestamps for new and modified files
+- Check for browser support via cookie and Chrome-check
 
 ## [0.0.2] - 2019-03-20
 - Temporary release for CI purpose